Add unique_rows to API and move to main nimare utils
